+++ Sources/KeyboardKit/Context/KeyboardContext.swift
@@ -1,0 +1,19 @@
+//
+//  KeyboardContext.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2020-06-15.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+/**
+ This protocol can be implemented by any classes that can be
+ used to provide the keyboard extension with contextual info.
+ */
+public protocol KeyboardContext: AnyObject {
+    
+    var keyboardType: KeyboardType { get set }
+}